requirements.txt
certifi==2018.8.24
chardet==3.0.4
Click==7.0
idna==2.6
pkg-resources==0.0.0
python-apt==1.8.4
requests==2.21.0
RPi.GPIO==0.7.0
six==1.12.0
ssh-import-id==5.7
urllib3==1.24.1
w1thermsensor==1.3.0
